+++ src/date.rs
@@ -11,8 +11,9 @@ use crate::{
 };
 use std::{
     fmt::Display,
+    ops::{Add, AddAssign, Sub, SubAssign},
     str::FromStr,
-    time::{SystemTime, UNIX_EPOCH},
+    time::{Duration, SystemTime, UNIX_EPOCH},
 };
 
 /// Date units for functions like [`Date::get`] or [`Date::apply`].
@@ -403,6 +404,47 @@ impl From<&DateTime> for Date {
         Self {
             days: value.as_days(),
         }
+    }
+}
+
+impl Add<Duration> for Date {
+    type Output = Self;
+
+    /// Performs the `+` operation.
+    ///
+    /// Only adds full days (`86 400` seconds) to [`Date`]. Any additional duration will be ignored.
+    fn add(self, rhs: Duration) -> Self::Output {
+        let days = self.days + (rhs.as_secs() / SECS_PER_DAY_U64) as i32;
+        Self { days }
+    }
+}
+
+impl Sub<Duration> for Date {
+    type Output = Self;
+
+    /// Performs the `-` operation.
+    ///
+    /// Only removes full days (`86 400` seconds) to [`Date`]. Any additional duration will be ignored.
+    fn sub(self, rhs: Duration) -> Self::Output {
+        let days = self.days - (rhs.as_secs() / SECS_PER_DAY_U64) as i32;
+        Self { days }
+    }
+}
+impl AddAssign<Duration> for Date {
+    /// Performs the `+=` operation.
+    ///
+    /// Only adds full days (`86 400` seconds) to [`Date`]. Any additional duration will be ignored.
+    fn add_assign(&mut self, rhs: Duration) {
+        *self = *self + rhs;
+    }
+}
+
+impl SubAssign<Duration> for Date {
+    /// Performs the `-=` operation.
+    ///
+    /// Only removes full days (`86 400` seconds) to [`Date`]. Any additional duration will be ignored.
+    fn sub_assign(&mut self, rhs: Duration) {
+        *self = *self - rhs;
     }
 }