+++ CHANGELOG.md
@@ -14,6 +14,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Time::clear_until`
 - Impl `From<Date>`, `From<&Date>`, `From<Time>` and `From<&Time>` for `DateTime`
 - Impl `From<DateTime>` and `From<&DateTime>` for `Date` and `Time`
+- Impl `Add<Duration>`, `AddAssign<Duration>`, `Sub<Duration>`, `AddAssign<Duration>` for `DateTime`, `Date` and `Time`
+- Impl `Add`, `AddAssign`, `Sub`, `AddAssign` for `Time`
+
+### Removed
+- `DateTime::date` in favor of `Date::from<DateTime>`
+- `DateTime::time` in favor of `Time::from<DateTime>`
 
 ## [0.3.0] - 2023-04-11
 ### Added
@@ -22,9 +28,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `DateTime::parse`
 - `Date::parse`
 - `Time::parse`
-### Removed
-- `DateTime::date` in favor of `Date::from<DateTime>`
-- `DateTime::time` in favor of `Time::from<DateTime>`
 
 ## [0.2.0] - 2022-08-25
 ### Added